Shifted the cart related functions to cart.js, and implemented the update Item function
